requirements.txt
click==8.1.3
comtypes==1.1.14
deprecation==2.1.0
docx2pdf==0.1.8
fpdf==1.7.2
img2pdf==0.4.4
importlib-metadata==5.0.0
lxml==4.9.1
packaging==21.3
pikepdf==6.2.2
Pillow==9.3.0
ppt2pdf==0.0.3
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.13.0
PyInquirer==1.0.3
pyparsing==3.0.9
PyPDF2==2.11.1
python-dotenv==0.21.0
regex==2022.10.31
six==1.16.0
tqdm==4.64.1
typing_extensions==4.4.0
wcwidth==0.2.5
zipp==3.10.0
